requirements.txt
httpx>=0.27.0
pyyaml>=6.0.1
pydantic>=2.6.0
python-dotenv>=1.0.1
py-clob-client>=0.34.0
